+++ packages/stream_chat_flutter/lib/src/components/stream_chat_component_builders.dart
@@ -13,7 +13,6 @@ Iterable<StreamComponentBuilderExtension<Object>> streamChatComponentBuilders({
   StreamComponentBuilder<MessageComposerInputHeaderProps>? messageComposerInputHeader,
   StreamComponentBuilder<MessageComposerInputTrailingProps>? messageComposerInputTrailing,
   StreamComponentBuilder<StreamMessageWidgetProps>? messageWidget,
-  StreamComponentBuilder<UnreadIndicatorProps>? unreadIndicator,
   StreamComponentBuilder<StreamMessageComposerAttachmentListProps>? messageComposerAttachmentList,
   StreamComponentBuilder<StreamMessageComposerAttachmentProps>? messageComposerAttachment,
   StreamComponentBuilder<StreamImageAttachmentProps>? imageAttachment,
@@ -24,6 +23,7 @@ Iterable<StreamComponentBuilderExtension<Object>> streamChatComponentBuilders({
   StreamComponentBuilder<StreamLinkPreviewAttachmentProps>? linkPreviewAttachment,
   StreamComponentBuilder<StreamVoiceRecordingAttachmentProps>? voiceRecordingAttachment,
   StreamComponentBuilder<StreamPollAttachmentProps>? pollAttachment,
+  StreamComponentBuilder<StreamUnsupportedAttachmentProps>? unsupportedAttachment,
 }) {
   final builders = [
     if (channelListItem != null) StreamComponentBuilderExtension(builder: channelListItem),
@@ -36,7 +36,6 @@ Iterable<StreamComponentBuilderExtension<Object>> streamChatComponentBuilders({
     if (messageComposerInputHeader != null) StreamComponentBuilderExtension(builder: messageComposerInputHeader),
     if (messageComposerInputTrailing != null) StreamComponentBuilderExtension(builder: messageComposerInputTrailing),
     if (messageWidget != null) StreamComponentBuilderExtension(builder: messageWidget),
-    if (unreadIndicator != null) StreamComponentBuilderExtension(builder: unreadIndicator),
     if (messageComposerAttachmentList != null) StreamComponentBuilderExtension(builder: messageComposerAttachmentList),
     if (messageComposerAttachment != null) StreamComponentBuilderExtension(builder: messageComposerAttachment),
     if (imageAttachment != null) StreamComponentBuilderExtension(builder: imageAttachment),
@@ -47,6 +46,7 @@ Iterable<StreamComponentBuilderExtension<Object>> streamChatComponentBuilders({
     if (linkPreviewAttachment != null) StreamComponentBuilderExtension(builder: linkPreviewAttachment),
     if (voiceRecordingAttachment != null) StreamComponentBuilderExtension(builder: voiceRecordingAttachment),
     if (pollAttachment != null) StreamComponentBuilderExtension(builder: pollAttachment),
+    if (unsupportedAttachment != null) StreamComponentBuilderExtension(builder: unsupportedAttachment),
   ];
 
   return builders;