+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ca.dart
@@ -725,4 +725,7 @@ class StreamChatLocalizationsCa extends GlobalStreamChatLocalizations {
 
   @override
   String get openFilesLabel => 'Obrir fitxers';
+
+  @override
+  String get unsupportedAttachmentLabel => 'Adjunt no compatible';
 }