+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_de.dart
@@ -723,4 +723,7 @@ class StreamChatLocalizationsDe extends GlobalStreamChatLocalizations {
 
   @override
   String get openFilesLabel => 'Dateien öffnen';
+
+  @override
+  String get unsupportedAttachmentLabel => 'Nicht unterstützter Anhang';
 }