+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -729,4 +729,7 @@ Limite de pièces jointes dépassée : il n'est pas possible d'ajouter plus de $
 
   @override
   String get openFilesLabel => 'Ouvrir des fichiers';
+
+  @override
+  String get unsupportedAttachmentLabel => 'Pièce jointe non prise en charge';
 }