+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_it.dart
@@ -732,4 +732,7 @@ Attenzione: il limite massimo di $limit file è stato superato.
 
   @override
   String get openFilesLabel => 'Apri file';
+
+  @override
+  String get unsupportedAttachmentLabel => 'Allegato non supportato';
 }