+++ packages/stream_chat_localizations/test/translations_test.dart
@@ -327,6 +327,7 @@ void main() {
       expect(localizations.openCameraLabel, isNotNull);
       expect(localizations.selectFilesToShareLabel, isNotNull);
       expect(localizations.openFilesLabel, isNotNull);
+      expect(localizations.unsupportedAttachmentLabel, isNotNull);
     });
   }